Tidy up before adding new schema
