+++ backend/app/db/models/auditable.py
@@ -1,0 +1,7 @@
+import sqlalchemy as sa
+from sqlalchemy.sql import func
+
+
+class Auditable:  # noqa: D101
+    created_ts = sa.Column(sa.DateTime(timezone=True), server_default=func.now())
+    updated_ts = sa.Column(sa.DateTime(timezone=True), onupdate=func.now())
